moved add list logic into TodoPage
