+++ src/components/Display/TodoPage/index.js
@@ -3,7 +3,7 @@
 // TodoPage Component - Contains NavBar, SideBar, and Todo 
 // components
 
-import React, { useState } from 'react';
+import React, { useState, useEffect, useRef } from 'react';
 import { Dialog, DialogContent, DialogActions, 
     Button, Snackbar, withStyles, IconButton } from '@material-ui/core';
 import {Close as CloseIcon} from '@material-ui/icons';
@@ -12,6 +12,8 @@ import Todo from '../Todo';
 import NavBar from '../NavBar';
 import SideBar from '../SideBar';
 import AddListForm from '../AddListForm';
+
+import * as Firestore from '../../Firestore';
 
 const styles = theme => ({
     navBar: {
@@ -38,12 +40,47 @@ function TodoPage(props) {
 
     const { classes } = props;
     // const [todoLists, setTodoLists] = useState({})
+    const isFirstRun = useRef(true);
+    const [todoListList, setTodoListList] = useState([]);
     const [addListOpen, setAddListOpen] = useState(false);
     const [addedSnackbarOpen, setAddedSnackbarOpen] = useState(false);
     const [addListError, setAddListError] = useState("");
     const [listToAddLocally, setListToAddLocally] = useState({id: -1});
+    const [selectedTodoListId, setSelectedTodoListId] = useState(-1)
     // will be used for loading symbol
     // const [addingList, setAddingList] = useState(false);
+
+    // run once on startup
+    useEffect(() => {
+        Firestore.getAllTodoLists().then((allLists) => {
+            let todoLists = [];
+            allLists.forEach(doc => {
+                let list = doc.data();
+                list.id = doc.id;
+                todoLists.push(list);
+            })
+            setTodoListList(todoLists);
+            // console.log(response);
+        })
+  
+    }, [])
+
+    useEffect(() => {
+        // don't run on initial load
+        if (isFirstRun.current) {
+          isFirstRun.current = false;
+          return;
+        }
+        // check if list to add is a real one
+        if (listToAddLocally.id !== -1) {
+          // get list of todo lists
+          let tempList = todoListList;
+          tempList.push(listToAddLocally); // add new one
+          setListToAddLocally({id: -1});  // reset list to add to dummy
+          setTodoListList(tempList);  // set new list of lists
+        }
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+      }, [listToAddLocally]);
 
     function handleAddListOpen() {
         setAddListOpen(true);
@@ -94,7 +131,8 @@ function TodoPage(props) {
                 <NavBar />
             </div>
             <div className={classes.sideBar}>
-                <SideBar handleAddListOpen={handleAddListOpen} listToAddLocally={listToAddLocally} setListToAddLocally={setListToAddLocally}/>
+                <SideBar todoListList={todoListList} handleAddListOpen={handleAddListOpen} 
+                listToAddLocally={listToAddLocally} setListToAddLocally={setListToAddLocally}/>
             </div>
             <Todo className={classes.todoMain}/>
             <Dialog border={2} open={addListOpen} aria-labelledby="form-dialog-title"
